+++ sensor_data/publish_sensor_readings.py
@@ -4,7 +4,7 @@ from rabbitmq.publisher import CSVRecordPublisher
 
 def main():
     parser = argparse.ArgumentParser(prog="publish_sensor_readings")
-    parser.add_argument("-T", "--period", type=int)
+    parser.add_argument("-T", "--period", type=float)
     args = parser.parse_args()
 
     publisher = CSVRecordPublisher(
